pyproject.toml
[tool.poetry]
name = "aplazame-backend"
version = "0.1.0"
description = "Aplazame Backend"
authors = ["Pablo Cabezas <pabcabsal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.13"
uvicorn = "^0.11"
jinja2 = "^2.11"
pyyaml = "^5.3"
ujson = "^3.0"
itsdangerous = "^1.1"
python-dotenv = "^0.14"
tortoise-orm = { extras = ["accel"], version = "^0.16.13" }
pydantic = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = { version = "^6.0.0-rc", allow-prereleases = true }
pytest-asyncio = "^0.14.0"
async-asgi-testclient = "^1.4.4"
Faker = "^4.1.1"
asynctest = "^0.13.0"
pytest-cov = "^2.10.0"
safety = "^1.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


